Change quantity type to integer

+++ services/order/src/main/java/dev/berke/app/product/PurchaseRequest.java
@@ -9,6 +9,6 @@ public record PurchaseRequest(
         Integer productId,
 
         @Positive(message = OrderConstants.QUANTITY_POSITIVE_MESSAGE)
-        double quantity
+        Integer quantity
 ) {
 }
