+++ services/product/src/main/java/dev/berke/app/product/ProductPurchaseRequest.java
@@ -4,11 +4,10 @@ import dev.berke.app.constants.ProductConstants;
 import jakarta.validation.constraints.NotNull;
 
 public record ProductPurchaseRequest(
-
         @NotNull(message = ProductConstants.PRODUCT_ID_MANDATORY_MESSAGE)
         Integer productId,
 
         @NotNull(message = ProductConstants.QUANTITY_MANDATORY_MESSAGE)
-        double quantity
+        Integer quantity
 ) {
 }
